parent end apply app changes

+++ app/client-end/src/eden/sun/childrenguard/child/receiver/PackageChangeReceiver.java
@@ -105,8 +105,10 @@ public class PackageChangeReceiver extends BroadcastReceiver{
 					ViewDTO<AppViewDTO> view = JSONUtil.getInstallAppView(response);
 			    	
 			    	if( view.getMsg().equals(ViewDTO.MSG_SUCCESS)){
-			    		AppDao appDao = new AppDao(context);
-			    		appDao.delete(view.getData().getId());
+			    		if( view.getData() != null ){
+			    			AppDao appDao = new AppDao(context);
+			    			appDao.delete(view.getData().getId());
+			    		}
 			    	}
 				}
